oppdatering

+++ tzuno/Strategy.h
@@ -3,6 +3,7 @@
 
 #include <Arduino.h>
 #include <WString.h>
+#include <Zumo
 
 /*
 Strategisuperklassen, alle strategier implementerer denne.
@@ -15,7 +16,8 @@ i valg av strategi
 
 class Strategy {
 protected:
-    int sonarDistance;
+    float sonarDistance;
+    
 public:
     // Konstruktør og destruktør gjør ingenting annet enn 
     // å måtte være implementert.
@@ -25,7 +27,7 @@ public:
     // Virtuelle funksjoner må implementeres av alle strategiklasser
     virtual void run() = 0;
 
-    void setSonarDistance(int d);
+    void setSonarDistance(float d);
 };
 
 #endif
